+++ R/write_aniframe.R
@@ -63,8 +63,8 @@ write_aniframe <- function(data, filename, ...) {
 write_aniframe_csv <- function(data, filename, ...) {
   dot_args <- list(...)
 
-  # # Validate filename
-  # ensure_file_has_expected_suffix(filename, "csv")
+  # Validate filename
+  ensure_file_has_expected_suffix(filename, c("csv", "tsv"))
 
   # Write data
   vroom::vroom_write(data, filename, ...) |>
@@ -76,19 +76,7 @@ write_aniframe_parquet <- function(data, filename, ...) {
   dot_args <- list(...)
 
   # Check that arrow is installed
-  rlang::check_installed(
-    "arrow",
-    reason = "to use write_aniframe()",
-    action = function(...) {
-      utils::install.packages(
-        'arrow',
-        repos = c(
-          'https://roaldarbol.r-universe.dev',
-          'https://cloud.r-project.org'
-        )
-      )
-    }
-  )
+  check_arrow()
 
   # Validate filename
   ensure_file_has_expected_suffix(filename, "parquet")