+++ NEWS.md
@@ -1,3 +1,9 @@
+# aniread 0.3.0
+
+* Adapt to the tidy movement data ethos implemented in *aniframe* 0.4.0.
+* Adds `read_trackmate` for TrackMate XML files. Thanks to @quantixed for writing the reader function in TrackMateR, which has been adapted here.
+* Adds `read_octron` for Octron CSV files.
+
 # aniread 0.2.0
 
 * Added a `NEWS.md` file to track changes to the package.